refactor(frontend): 🎨 consistent feature structure
