+++ app/frontend/src/features/auth/components/index.tsx
@@ -1,0 +1,2 @@
+export * from "./github-signin-button";
+export * from "./signout-button";
