+++ app/frontend/src/features/auth/hooks/index.tsx
@@ -1,0 +1,1 @@
+export * from "./use-user";
