--- app/frontend/src/features/auth/index.ts
@@ -1,3 +1,0 @@
-export * from "./context/auth-provider";
-export * from "./hooks/use-user";
-export * from "./queries";
