+++ app/frontend/src/features/profile/components/profile-edit-form.tsx
@@ -2,7 +2,7 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Separator } from "@/components/ui/separator";
-import { useUser } from "@/features/auth";
+import { useUser } from "@/features/auth/hooks";
 import {
 	useDeleteAvatarMutation,
 	useUpdateProfileMutation,
