+++ app/frontend/src/features/uploadthing/hooks/index.tsx
@@ -1,5 +1,4 @@
 import type { FileRouter } from "@server/api/uploadthing";
 import { generateReactHelpers } from "@uploadthing/react";
 
-export const { useUploadThing } =
-	generateReactHelpers<FileRouter>();
+export const { useUploadThing } = generateReactHelpers<FileRouter>();
